+++ replicante/src/components/discovery/worker.rs
@@ -1,4 +1,7 @@
+use std::collections::HashMap;
+use std::sync::Mutex;
 use std::time::Duration;
+
 use error_chain::ChainedError;
 use prometheus::Registry;
 use slog::Logger;
@@ -8,6 +11,7 @@ use replicante_agent_discovery::discover;
 
 use replicante_data_aggregator::Aggregator;
 use replicante_data_fetcher::Fetcher;
+use replicante_data_fetcher::Snapshotter;
 use replicante_data_models::ClusterDiscovery;
 use replicante_data_models::Event;
 
@@ -16,7 +20,10 @@ use replicante_streams_events::EventsStream;
 
 use super::super::super::Result;
 use super::super::super::ResultExt;
+use super::EventsSnapshotsConfig;
+
 use super::metrics::DISCOVERY_FETCH_ERRORS_COUNT;
+use super::metrics::DISCOVERY_SNAPSHOT_TRACKER_COUNT;
 
 
 const FAIL_FIND_DISCOVERY: &str = "Failed to fetch cluster discovery";
@@ -25,7 +32,8 @@ const FAIL_PERSIST_DISCOVERY: &str = "Failed to persist cluster discovery";
 
 /// Implements the discovery logic of a signle discovery loop.
 pub struct DiscoveryWorker {
-    config: BackendsConfig,
+    discovery_config: BackendsConfig,
+    emissions: EmissionTracker,
     events: EventsStream,
     logger: Logger,
     store: Store,
@@ -38,13 +46,15 @@ pub struct DiscoveryWorker {
 impl DiscoveryWorker {
     /// Creates a discover worker.
     pub fn new(
-        config: BackendsConfig, logger: Logger, events: EventsStream, store: Store,
-        timeout: Duration
+        discovery_config: BackendsConfig, snapshots_config: EventsSnapshotsConfig,
+        logger: Logger, events: EventsStream, store: Store, timeout: Duration
     ) -> DiscoveryWorker {
         let aggregator = Aggregator::new(logger.clone(), store.clone());
         let fetcher = Fetcher::new(logger.clone(), events.clone(), store.clone(), timeout);
+        let emissions = EmissionTracker::new(snapshots_config);
         DiscoveryWorker {
-            config,
+            discovery_config,
+            emissions,
             events,
             logger,
             store,
@@ -63,7 +73,7 @@ impl DiscoveryWorker {
     /// Runs a signle discovery loop.
     pub fn run(&self) {
         debug!(self.logger, "Discovering agents ...");
-        for cluster in discover(self.config.clone()) {
+        for cluster in discover(self.discovery_config.clone()) {
             match cluster {
                 Ok(cluster) => self.process(cluster),
                 Err(error) => {
@@ -89,7 +99,8 @@ impl DiscoveryWorker {
     ///   5. Pass the discovery to the status aggregator (TODO: move when tasks are in place).
     fn process(&self, cluster: ClusterDiscovery) {
         let name = cluster.cluster.clone();
-        if let Err(error) = self.process_checked(cluster) {
+        let snapshot = self.emissions.snapshot(name.clone());
+        if let Err(error) = self.process_checked(cluster, snapshot) {
             let error = error.display_chain().to_string();
             error!(
                 self.logger, "Failed to process cluster discovery";
@@ -99,7 +110,20 @@ impl DiscoveryWorker {
         }
     }
 
-    fn process_checked(&self, cluster: ClusterDiscovery) -> Result<()> {
+    fn process_checked(&self, cluster: ClusterDiscovery, snapshot: bool) -> Result<()> {
+        let name = cluster.cluster.clone();
+        if snapshot {
+            let snapshotter = Snapshotter::new(
+                name.clone(), self.events.clone(), self.store.clone()
+            );
+            if let Err(error) = snapshotter.run() {
+                let error = error.display_chain().to_string();
+                error!(
+                    self.logger, "Failed to emit snapshots";
+                    "cluster" => name, "error" => error
+                );
+            }
+        }
         self.process_discovery(cluster.clone())?;
         self.fetcher.process(cluster.clone());
         self.aggregator.process(cluster);
@@ -129,5 +153,46 @@ impl DiscoveryWorker {
         let event = Event::builder().cluster().cluster_new(cluster.clone());
         self.events.emit(event).chain_err(|| FAIL_PERSIST_DISCOVERY)?;
         self.store.persist_discovery(cluster).chain_err(|| FAIL_PERSIST_DISCOVERY)
+    }
+}
+
+
+/// Helper object to decide when snapshot events should be emitted.
+///
+/// This is a naive implementation using an in-memory map that never forgets nodes.
+/// As a consequence it can "leak" memory in the presence of frequent cluster rotation.
+struct EmissionTracker {
+    enabled: bool,
+    frequency: u32,
+    state: Mutex<HashMap<String, u32>>,
+}
+
+impl EmissionTracker {
+    pub fn new(config: EventsSnapshotsConfig) -> EmissionTracker {
+        EmissionTracker {
+            enabled: config.enabled,
+            frequency: config.frequency,
+            state: Mutex::new(HashMap::new()),
+        }
+    }
+
+    /// Determine if it is time to snapshot a cluster.
+    pub fn snapshot(&self, cluster: String) -> bool {
+        if !self.enabled {
+            return false;
+        }
+        let mut map = self.state.lock().expect("EmissionTracker lock was poisoned");
+        // Default to 1 so that we can emit immediatelly.
+        // This is so that a failover leads to a double snapshot instead of a snapshot delay.
+        let state = map.entry(cluster).or_insert_with(|| {
+            DISCOVERY_SNAPSHOT_TRACKER_COUNT.inc();
+            1
+        });
+        *state -= 1;
+        if *state == 0 {
+            *state = self.frequency;
+            return true;
+        }
+        false
     }
 }